requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
ansiwrap==0.8.4
appnope==0.1.2
APScheduler==3.8.1
asttokens==2.0.5
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
bech32==1.2.0
bitarray==1.2.2
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
cytoolz==0.11.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.17.0
entrypoints==0.4
eth-abi==3.0.0
eth-account==0.6.1
eth-hash==0.3.2
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.0.0
eth-utils==2.0.0
executing==0.8.2
frozenlist==1.3.0
gmail==0.6.3
greenlet==1.1.2
grpcio==1.42.0
grpcio-tools==1.42.0
hdwallets==0.1.2
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mnemonic==0.20
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.11
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.17.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycryptodome==3.14.1
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rlp==3.0.0
six==1.16.0
SQLAlchemy==1.4.35
stack-data==0.2.0
tenacity==8.0.1
testpath==0.5.0
textwrap3==0.9.2
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.54.0
websockets==8.1
yarl==1.7.2
zipp==3.8.0
py-hs-openapi==1.1.2
